Beginning Auto-scan function + Graphical Inventory
David began working on an Auto-scan function for the robot as well as a way to delete blocks out of bound, I began working on a Graphical interface for the Robot's inventory. It can fetch the Robot's inventory and show information about it, has no way of moving items around yet.